[sparsity] Fix GPU training for sparsity
The GPU training was not supported in the sparsifier.
The reason was that when the sparsifier was created the masks would default to the CPU.
Attaching a GPU model to the sparsifier would throw an error.
The solution is to create the masks on the same device as the weight.

ghstack-source-id: 9eae9c1ee3dd09ad1b92b1e0a99e53f318be7a56
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66412